+++ src/main/java/com/spring/pro4/service/ContactService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ContactService {
     List<Contact> findAll();
     List<Contact> findByFirstName(String firstName);
-    Contact findByid(Long id);
+    Contact findByid(Integer id);
     Contact save(Contact contact);
     void delete(Contact contact);
 }
